requirements.txt
pandas
requests
python-dotenv
fastapi
uvicorn[standard]
apscheduler
aiofiles
